@@ -30,34 +30,36 @@
         ]
     },
     "require": {
+        "php": "^8.0",
         "ext-json": "*",
         "ext-pdo": "*",
-        "symfony/http-foundation": "^4.3 || ^5.0",
+        "symfony/http-foundation": "^6.0",
         "doctrine/orm": "^2.6",
-        "symfony/validator": "^4.3 || ^5.0",
-        "symfony/http-kernel": "^4.3 || ^5.0",
-        "symfony/dependency-injection": "^4.3 || ^5.0",
-        "symfony/config": "^4.3 || ^5.0",
-        "symfony/framework-bundle": "^4.3 || ^5.0",
-        "symfony/monolog-bridge": "^4.3 || ^5.0",
-        "symfony/event-dispatcher": "^4.3 || ^5.0",
-        "symfony/translation": "^4.3 || ^5.0",
+        "symfony/validator": "^6.0",
+        "symfony/http-kernel": "^6.0",
+        "symfony/dependency-injection": "^6.0",
+        "symfony/config": "^6.0",
+        "symfony/framework-bundle": "^6.0",
+        "symfony/monolog-bridge": "^6.0",
+        "symfony/event-dispatcher": "^6.0",
+        "symfony/translation": "^6.0",
         "symfony/monolog-bundle": "^3.2",
-        "symfony/security-core": "^4.3 || ^5.0",
-        "symfony/finder": "^4.3 || ^5.0",
-        "symfony/filesystem": "^4.3 || ^5.0",
-        "symfony/console": "^4.3 || ^5.0"
+        "symfony/security-core": "^6.0",
+        "symfony/finder": "^6.0",
+        "symfony/filesystem": "^6.0",
+        "symfony/console": "^6.0",
+        "symfony/security-http": "^6.0"
     },
     "require-dev": {
-        "phpunit/phpunit": "^8.0",
-        "symfony/phpunit-bridge": "^4.3 || ^5.0",
-        "symfony/cache": "^4.3 || ^5.0",
-        "symfony/debug": "^4.3 || ^5.0"
+        "phpunit/phpunit": "^9.0",
+        "symfony/phpunit-bridge": "^6.0",
+        "symfony/cache": "^6.0",
+        "symfony/debug": "^4.0"
     },
     "extra": {
         "symfony": {
             "allow-contrib": false,
-            "require": "^4.3 || ^5.0"
+            "require": "^6.0"
         }
     }
 }
