@@ -3,7 +3,7 @@
 /**
  * OpeniumSymfonyToolKit bundle
  *
- * PHP Version >=7.1
+ * PHP Version >=8.0
  *
  * @package  Openium\SymfonyToolKitBundle
  * @author   Openium <contact@openium.fr>
